Add transaction passthrough for wrapped documents
