+++ packages/paper/src/TODO.md
@@ -2,7 +2,6 @@
 
 ## TODO
 
-- [ ] Passthrough proxywrapped documents on mutation returns
 - [ ] (create issue) Hooks - auto ID hook (if ID has not been set, create a hash or auto-increment)
 - [ ] (create issue) validator `enable`/`disable` or `add`/`remove` options on `Paper`
 - [ ] (create issue) Create documents that do not correspond with an GraphQL type (add tests demonstrating how this would work)
@@ -41,4 +40,4 @@
 - [X] Hooks - before/after transaction
 - [X] Collapse validators on Paper to `validators.field` and `validators.document`
 - [X] Hooks - add `queueEvent` to hooks and transactions
-
+- [X] Passthrough proxywrapped documents on mutation returns
